[K8S] Fix a bug where k8s api exception are not converted to their status code dependent specific exception class Also fix and relocate k8s to the correct file.
